+++ src/main/java/kr/co/automl/domain/metadata/dataset/ContactPoint.java
@@ -1,4 +1,4 @@
-package kr.co.automl.domain.metadata;
+package kr.co.automl.domain.metadata.dataset;
 
 import java.util.Objects;
