+++ src/main/java/kr/co/automl/domain/metadata/dataset/Organization.java
@@ -1,4 +1,4 @@
-package kr.co.automl.domain.metadata;
+package kr.co.automl.domain.metadata.dataset;
 
 /**
  * 구축기관
